# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Echt HR & Process Automation Platform API
  version: 1.0.0
  description: REST API for HRMS Backend
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    MeResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
...
    AuditLogsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        total:
          type: integer
    LeaveRequest:
      type: object
      required:
        - leaveTypeId
        - startDate
        - endDate
        - workingDays
      properties:
        leaveTypeId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        workingDays:
          type: integer
        reason:
          type: string
    LeaveResponse:
      type: object
      properties:
        leave:
          type: object
        workflowId:
          type: string
          format: uuid
...
    ClockSyncBatch:
      type: array
      items:
        $ref: '#/components/schemas/ClockEvent'
      maxItems: 50
...
paths:
...
  /me:
    get:
      summary: Get current user profile
      description: Token claims + resolved roles
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/logout:
    post:
      summary: Logout user
      description: Invalidate current session (client-side token clearing)
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/azure/login:
    get:
      summary: Initiate Azure AD login
      description: Redirect to Microsoft Azure AD for authentication
      parameters:
        - name: redirect_uri
          in: query
          description: Where to redirect after authentication
          schema:
            type: string
            format: uri
            example: "http://localhost:3000/auth/callback"
      security: []
      responses:
        '302':
          description: Redirect to Azure AD
  /auth/azure/callback:
    get:
      summary: Azure AD callback
      description: Handle Azure AD authentication callback and exchange code for tokens
      parameters:
        - name: code
          in: query
          description: Authorization code from Azure AD
          schema:
            type: string
        - name: redirect_uri
          in: query
          description: Original redirect URI
          schema:
            type: string
            format: uri
      security: []
      responses:
        '302':
          description: Redirect to frontend with tokens
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Token exchange failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /audit/logs:
    get:
      summary: Get audit logs
      description: Paginated immutable audit trail
      parameters:
        - name: resourceType
          in: query
          schema:
            type: string
        - name: resourceId
          in: query
          schema:
            type: string
        - name: actorUserId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: take
          in: query
          schema:
            type: integer
        - name: skip
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /hr/leave:
    post:
      summary: Create leave request
      description: Create leave + workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequest'
      responses:
        '201':
          description: Leave created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List my leaves
      description: List caller's leave rows
      responses:
        '200':
          description: Leaves
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /attendance/clock:
    post:
      summary: Clock in/out
      description: Single clock-in/out (online)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockEvent'
      responses:
        '200':
          description: Clocked
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /attendance/clock/sync:
    post:
      summary: Batch clock sync
      description: Batch offline sync (max 50)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockSyncBatch'
      responses:
        '200':
          description: Synced
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /attendance/users/{userId}/clock-events:
    get:
      summary: Get clock events for user
      description: Location audit (lat/lng, distance)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Clock events
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /reports/attendance:
    get:
      summary: Get attendance report
      tags: [Reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Attendance report
  /reports/leave:
    get:
      summary: Get leave report
      tags: [Reports]
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Leave report
  /reports/payroll:
    get:
      summary: Get payroll report
      tags: [Reports]
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Payroll report
  /leave:
    get:
      summary: Get all leave requests
      tags: [Leave]
      parameters:
        - name: employee_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Leave requests
    post:
      summary: Create leave request
      tags: [Leave]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaveRequest'
      responses:
        '201':
          description: Leave created
...
  /leave/{id}:
    put:
      summary: Update leave request
      tags: [Leave]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Leave updated
    delete:
      summary: Delete leave request
      tags: [Leave]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Leave deleted
  /leave/{id}/approve:
    put:
      summary: Approve leave request
      tags: [Leave]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Leave approved
  /leave/{id}/reject:
    put:
      summary: Reject leave request
      tags: [Leave]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Leave rejected
  /leave/{id}/return:
    put:
      summary: Mark leave as returned
      tags: [Leave]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Leave returned
  /leave/{id}/submit:
    put:
      summary: Submit leave request
      tags: [Leave]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Leave submitted
  /attendance:
    get:
      summary: Get all attendance records
      tags: [Attendance]
      parameters:
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attendance records
...
  /attendance/corrections:
    get:
      summary: Get attendance corrections
      tags: [Attendance]
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Corrections
    post:
      summary: Request attendance correction
      tags: [Attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceCorrection'
      responses:
        '201':
          description: Correction requested
  /attendance/checkin:
    post:
      summary: Check in
      tags: [Attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockEvent'
      responses:
        '200':
          description: Checked in
  /attendance/checkout:
    post:
      summary: Check out
      tags: [Attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockEvent'
      responses:
        '200':
          description: Checked out
  /attendance/corrections/{id}/approve:
    put:
      summary: Approve attendance correction
      tags: [Attendance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Correction approved
  /hr/birthdays/upcoming:
    get:
      summary: Get upcoming birthdays
      tags: [HR]
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Upcoming birthdays
...